+++ src/main/java/com/koreait/movie/command/my_page/My_writeCommand.java
@@ -1,20 +1,18 @@
 package com.koreait.movie.command.my_page;
 
-import java.util.Map;
-
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.ui.Model;
 
-import com.koreait.movie.common.CommonMapCommand;
+import com.koreait.movie.common.CommonVoidCommand;
+import com.koreait.movie.dao.My_dao;
 
-public class My_writeCommand implements CommonMapCommand {
+public class My_writeCommand implements CommonVoidCommand {
 
 	@Override
-	public Map<String, Object> execute(SqlSession sqlSession, Model model) {
-
+	public void execute(SqlSession sqlSession, Model model) {
 		
-		
-		return null;
+		My_dao my_dao = sqlSession.getMapper(My_dao.class);
+		model.addAttribute("my_write", my_dao.my_write_page());
 	}
 
 }
