--- src/main/java/com/koreait/movie/dao/Comment_dao.java
@@ -1,13 +1,0 @@
-package com.koreait.movie.dao;
-
-import com.koreait.movie.dto.CommentDto;
-
-public class Comment_dao {
-	
-	public CommentDto view(int no) {
-			
-		return null;
-	
-		}
-
-}
