+++ src/main/java/com/koreait/movie/dao/Movie_info_dao.java
@@ -1,6 +1,15 @@
 package com.koreait.movie.dao;
 
+import com.koreait.movie.dto.CommentDto;
+
 public interface Movie_info_dao {
+	
+	public Integer movie_no(int movie_no);
+	
+	public CommentDto my_write(int user_no);
+	
+	public CommentDto view(int no);
+	
 	
 
 }
