--- src/main/java/com/koreait/movie/dao/My_write_dao.java
@@ -1,15 +1,0 @@
-package com.koreait.movie.dao;
-
-import java.sql.Date;
-
-public interface My_write_dao {
-	
-	public Integer user_no(int user_no);
-	
-	public Integer movie_no(int movie_no);
-	
-	public String comments_title(String comments_title);
-	
-	public Date comment_date(Date comment_date); 
-
-}
