--- src/main/java/com/koreait/movie/dao/My_write_view_dao.java
@@ -1,9 +1,0 @@
-package com.koreait.movie.dao;
-
-import com.koreait.movie.dto.CommentDto;
-
-public interface My_write_view_dao {
-	
-	public CommentDto my_write(int user_no);
-
-}
