--- src/main/java/com/koreait/movie/dao/mapper/comment_mapper.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper 
-PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
-"http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-
-
-<mapper namespace="com.koreait.movie.dao.Comment_dao">
-	
-	<select id="CommentDto" resultType="com.koreait.movie.dto.CommentDto">
-		SELECT movie_no, comments_title, comment_date
-		  FROM comment
-		  where user_no = #{user_no}
-	</select>
-	
-	<select id="view" parameterType="int" resultType="com.koreait.movie.dto.CommentDto">
-    	Select * 
-    		from comment 
-    		where user_no = #{user_no}
-	</select>
-	
-	<!-- 최신 NO가 위로 올라오는 순서로 정렬한 테이블 :
-	<select id="comment_list" resultType="com.koreait.movie.dto.CommentDto">
-		SELECT B.*
-		  FROM (SELECT ROWNUM AS RN, A.*
-		          FROM (SELECT *
-		                  FROM comment
-		              ORDER BY NO DESC) A ) B
-		 WHERE B.RN BETWEEN #{param1} AND #{param2}
-	</select>
-		 -->
-	
-</mapper>
