--- src/main/java/com/koreait/movie/dao/mapper/my_write_mapper.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper 
-PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
-"http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-
-
-<mapper namespace="com.koreait.movie.dao.My_write_dao">
-	
-	<select id="my_write" parameterType="int" resultType="com.koreait.movie.dto.CommentDto">
-		SELECT user_no, movie_no, comments_title, comment_date
-		  FROM comments
-		  where user_no = #{user_no}
-	</select>
-	
-	
-	
-	
-	
-	
-</mapper>
